refactor: extract development timeline into own component
now we can reuse it elsewhere, e.g., the requirement details page
