+++ src/components/RequirementCard.vue
@@ -17,24 +17,11 @@
     </template>
 
     <template #content>
-    <vue3-markdown-it :source="description" class="p-mt-3 p-mb-3" />
-    <!-- Timeline -->
-    <div v-if="issue_url">
-      <div class="p-d-flex p-jc-start p-ai-center">
-        <h2>{{ t('devTimeline-title') }}</h2>
-        <a v-if="issue_url" :href="issue_url" target="_blank" rel="noreferrer" class="p-ml-2"><i class="pi pi-github"></i> {{ t('viewOnGitHub') }}</a>
-      </div>
-      <Timeline :value="timelineEvents" layout="horizontal" align="bottom">
-          <template #marker="slotProps">
-            <span class="custom-marker">
-              <i :class="slotProps.item.status"></i>
-            </span>
-          </template>
-          <template #content="slotProps">
-            {{slotProps.item.label}}
-          </template>
-      </Timeline>
-    </div>
+      <vue3-markdown-it :source="description" class="p-mt-3 p-mb-3" />
+
+      <!-- Timeline -->
+      <RequirementDevTimeline :additionalProperties="additionalProperties" />
+
       <Dialog :header="t('editRequirement')" v-model:visible="displayRequirementEditor" :breakpoints="{'960px': '75vw', '640px': '100vw'}" :style="{width: '50vw'}" :modal="true">
         <RequirementEditor
           class="requirementEditor"
@@ -74,6 +61,7 @@ import { useI18n } from 'vue-i18n';
 import { ActionTypes } from '../store/actions';
 import { useConfirm } from "primevue/useconfirm";
 import CommentsList from './CommentsList.vue';
+import RequirementDevTimeline from '@/components/RequirementDevTimeline.vue';
 import { useRoute } from 'vue-router';
 
 import { routePathToRequirement } from '@/router';
@@ -82,7 +70,7 @@ import RequirementEditor from '../components/RequirementEditor.vue';
 import { getEnabledCategories } from 'trace_events';
 
 export default defineComponent({
-  components: { CommentsList, RequirementEditor },
+  components: { CommentsList, RequirementEditor, RequirementDevTimeline },
   name: 'RequirementCard',
   props: {
     id: { type: Number, required: true },
@@ -136,57 +124,6 @@ export default defineComponent({
         return additionalProperties.value.issue_url
       }
     });
-
-    // get issue's additionalProperties Descriptors & Values
-    const additionalPropertiesValue = additionalProperties.value;
-
-    if(additionalPropertiesValue === undefined || additionalPropertiesValue === null){
-      var issueNumberValue = undefined;
-      var issueStatusValue = undefined;
-
-    }else{
-          const issueNumberDescriptor = Object.getOwnPropertyDescriptor(additionalProperties.value, 'issue_number');
-          issueNumberValue = issueNumberDescriptor?.value;
-          const issueStatusDescriptor = Object.getOwnPropertyDescriptor(additionalProperties.value, 'issue_status');
-          issueStatusValue = issueStatusDescriptor?.value;
-    }
-
-    // **** Functions to update the timeline icons *****
-    // Requirement exported
-    function timelineIssueNumber(){
-      let timelineIssueNumberIcon = "pi pi-circle-off";
-      if(issueNumberValue !== undefined){
-        timelineIssueNumberIcon = "pi pi-check-circle";
-      }
-      return timelineIssueNumberIcon;
-    };
-    // Requirement opened or closed
-    function timelineIssueStatus(){
-      let timelineIssueStatusIcon = "pi pi-circle-off";
-      if(issueStatusValue !== undefined){
-        (issueStatusValue == "closed")
-          ? timelineIssueStatusIcon = 'pi pi-check-circle'
-          : (issueStatusValue == "opened")
-            ? timelineIssueStatusIcon = 'pi pi-circle-off'
-            : null;
-      }
-      return timelineIssueStatusIcon;
-    };
-    // Requirement Url (Development in Progress)
-    function timelineIssueUrl(){
-      let timelineIssueUrlIcon = "pi pi-circle-off";
-      if(issue_url.value != undefined){
-        timelineIssueUrlIcon = "pi pi-check-circle";
-      }
-      return timelineIssueUrlIcon;
-    };
-
-    //Timeline events
-    const timelineEvents = ref([
-      {label: t('devTimeline-requirementExported'), status: timelineIssueNumber()},
-      {label: t('devTimeline-inProgress'), status: timelineIssueUrl()},
-      {label: t('devTimeline-closed'), status: timelineIssueStatus()},
-      ]);
 
     const alertLogin = (message: string) => {
       confirm.require({
@@ -256,7 +193,7 @@ export default defineComponent({
       const requirementTitle = name.value.replace(/\s/g, '+');
       const requirementDescription = description.value.replace(/\s/g, '+');
       const bazaarRequirementUrl = createShareableRequirementLink()
-      if(additionalPropertiesValue == undefined){
+      if(additionalProperties.value == undefined){
         confirm.require({
           header: t('createIssueDialogHeader'),
           message: t('createIssueDialogMessage'),
@@ -388,7 +325,6 @@ export default defineComponent({
       displayRequirementEditor,
       requirementEditorCanceled,
       requirementEditorSaved,
-      timelineEvents,
       issue_url,
     };
   },
